refactor(ui): extract editor table styles and controls
Move table CSS into tiptap-node/table-node and the table toolbar into
tiptap-ui/table-controls, matching the convention used by every other
node and toolbar button. Drop two dead rules: .tableWrapper, whose
element only exists when column resizing is enabled, and a leftover
markdown-table-row selector nothing sets. Add the missing .selectedCell
rule so multi-cell selection is visible. No behavior change.
