+++ frontend/src/components/tiptap-templates/simple/simple-editor.scss
@@ -167,43 +167,3 @@
 .simple-editor-content .tiptap.ProseMirror.simple-editor ::selection {
   background: hsl(var(--primary) / 0.15);
 }
-
-.simple-editor-content
-  .tiptap.ProseMirror.simple-editor
-  p[data-markdown-table-row="true"] {
-  font-family: ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas,
-    "Liberation Mono", "Courier New", monospace;
-  white-space: pre;
-  margin: 0;
-}
-
-.simple-editor-content .tiptap.ProseMirror.simple-editor .tableWrapper {
-  margin: 0.75rem 0;
-  max-width: 100%;
-  overflow-x: auto;
-}
-
-.simple-editor-content .tiptap.ProseMirror.simple-editor table {
-  width: 100%;
-  border-collapse: collapse;
-  border-spacing: 0;
-  table-layout: auto;
-  font-variant-numeric: tabular-nums;
-}
-
-.simple-editor-content .tiptap.ProseMirror.simple-editor th,
-.simple-editor-content .tiptap.ProseMirror.simple-editor td {
-  border: 1px solid hsl(var(--border));
-  padding: 0.375rem 0.5rem;
-  text-align: left;
-  vertical-align: top;
-}
-
-.simple-editor-content .tiptap.ProseMirror.simple-editor th {
-  font-weight: 600;
-  background-color: hsl(var(--muted) / 0.35);
-}
-
-.simple-editor-content .tiptap.ProseMirror.simple-editor tbody tr:nth-of-type(odd) td {
-  background-color: hsl(var(--muted) / 0.18);
-}
