+++ frontend/src/components/tiptap-ui/table-controls/index.tsx
@@ -1,0 +1,2 @@
+export * from "./table-controls"
+export * from "./use-table-controls"
